requirements.txt
Django
djangorestframework
django-ckeditor
Pillow
djangorestframework 
dj-rest-auth 
django-allauth 
djangorestframework-simplejwt
requests
cryptography
